% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.glacier_package.R
\docType{package}
\name{paws.glacier-package}
\alias{paws.glacier}
\alias{paws.glacier-package}
\title{paws.glacier: Amazon Glacier}
\description{
Amazon Glacier is a storage solution for \"cold data.\"
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
