% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.glacier_operations.R
\name{purchase_provisioned_capacity}
\alias{purchase_provisioned_capacity}
\title{This operation purchases a provisioned capacity unit for an AWS account}
\usage{
purchase_provisioned_capacity(accountId)
}
\arguments{
\item{accountId}{[required] The AWS account ID of the account that owns the vault. You can either specify an AWS account ID or optionally a single '-' (hyphen), in which case Amazon Glacier uses the AWS account ID associated with the credentials used to sign the request. If you use an account ID, don't include any hyphens ('-') in the ID.}
}
\description{
This operation purchases a provisioned capacity unit for an AWS account.
}
\section{Accepted Parameters}{
\preformatted{purchase_provisioned_capacity(
  accountId = "string"
)
}
}

\examples{
# The example purchases provisioned capacity unit for an AWS account.
\donttest{purchase_provisioned_capacity(
  accountId = "-"
)}

}
