% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.glacier_operations.R
\name{get_data_retrieval_policy}
\alias{get_data_retrieval_policy}
\title{This operation returns the current data retrieval policy for the account and region specified in the GET request}
\usage{
get_data_retrieval_policy(accountId)
}
\arguments{
\item{accountId}{[required] The \code{AccountId} value is the AWS account ID. This value must match the AWS account ID associated with the credentials used to sign the request. You can either specify an AWS account ID or optionally a single '\code{-}' (hyphen), in which case Amazon Glacier uses the AWS account ID associated with the credentials used to sign the request. If you specify your account ID, do not include any hyphens ('-') in the ID.}
}
\description{
This operation returns the current data retrieval policy for the account and region specified in the GET request. For more information about data retrieval policies, see \href{http://docs.aws.amazon.com/amazonglacier/latest/dev/data-retrieval-policy.html}{Amazon Glacier Data Retrieval Policies}.
}
\section{Accepted Parameters}{
\preformatted{get_data_retrieval_policy(
  accountId = "string"
)
}
}

\examples{
# The example returns the current data retrieval policy for the account.
\donttest{get_data_retrieval_policy(
  accountId = "-"
)}

}
