% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.glacier_package.R
\docType{package}
\name{paws.glacier-package}
\alias{paws.glacier}
\alias{paws.glacier-package}
\title{paws.glacier: Amazon Glacier}
\description{
Amazon Glacier is an extremely low-cost storage
service that provides secure, durable, and easy-to-use storage for
data backup and archival. With Amazon Glacier, customers can store
their data cost effectively for months, years, or decades. Amazon
Glacier also enables customers to offload the administrative burdens
of operating and scaling storage to AWS, so they don\'t have to worry
about capacity planning, hardware provisioning, data replication,
hardware failure and recovery, or time-consuming hardware migrations.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker
}

}
\keyword{internal}
